Fix int32 overflow in SM89 paged MQA logits block addressing
block_idx loads as int32 and block_idx * stride_kvb wraps past 2^31 on
large unified KV pools (TP=8, >2k blocks), producing a negative offset
and an illegal memory access on long contexts / prefix caching.
Cast to int64 before the multiply, matching upstream 5201cdf4595f.

+++ vllm/models/deepseek_v4/nvidia/ops/sm12x_mqa.py
@@ -272,7 +272,7 @@ def _fp8_paged_mqa_logits_kernel(
         + block_rank[None, :] * stride_btk,
         mask=valid_m[:, None] & valid_n[None, :],
         other=0,
-    )
+    ).to(tl.int64)
 
     logits = tl.zeros((BLOCK_M, BLOCK_N), dtype=tl.float32)
     scale = tl.load(
@@ -397,7 +397,7 @@ def _fp8_paged_mqa_logits_rowwise_kernel(
         block_tables_ptr + batch * stride_btb + block_rank * stride_btk,
         mask=valid_row & context_mask,
         other=0,
-    )
+    ).to(tl.int64)
 
     scale = tl.load(
         scale_ptr + block_idx * stride_sb + block_offset * stride_ss,
